@@ -15,26 +15,37 @@
     "issues": "https://github.com/thefrosty/wp-upgrade-task-runner/issues"
   },
   "config": {
-    "sort-packages": true,
-    "optimize-autoloader": true
+    "allow-plugins": {
+      "composer/installers": true,
+      "dealerdirect/phpcodesniffer-composer-installer": true,
+      "roots/wordpress-core-installer": true
+    },
+    "optimize-autoloader": true,
+    "platform": {
+      "php": "7.4.26"
+    },
+    "process-timeout": 600,
+    "sort-packages": true
   },
   "minimum-stability": "dev",
   "prefer-stable": true,
   "require": {
-    "php": "^7.4|^8.0",
+    "php": "^7.4 || ^8.0",
     "pimple/pimple": "^3.2",
-    "symfony/http-foundation": "^4.4.7|^5.0.7",
-    "thefrosty/wp-utilities": "^2.0"
+    "symfony/http-foundation": "^5.2 || ~6.0",
+    "thefrosty/wp-utilities": "^2.5.1"
   },
   "require-dev": {
     "dealerdirect/phpcodesniffer-composer-installer": "^0.7.0",
     "phpcompatibility/php-compatibility": "^9.3",
-    "phpunit/phpunit": "^6|^7",
+    "phpunit/phpunit": "^9",
+    "roots/wordpress": "~5.9.1",
     "slevomat/coding-standard": "^4",
     "squizlabs/php_codesniffer": "^3.2",
     "wp-cli/wp-cli-bundle": "^2.4",
     "wp-coding-standards/wpcs": "^2.3",
-    "wp-phpunit/wp-phpunit": "^5.4"
+    "wp-phpunit/wp-phpunit": "~5.9.1",
+    "yoast/phpunit-polyfills": "^1.0.1"
   },
   "autoload": {
     "psr-4": {
