@@ -6,11 +6,11 @@
  * Plugin Name: Upgrade Task Runner
  * Plugin URI: https://github.com/thefrosty/wp-upgrade-task-runner
  * Description: A WordPress plugin for developers to write custom migration tasks.
- * Version: 2.5.1
+ * Version: 2.6.0
  * Author: Austin Passy
  * Author URI: https://github.com/thefrosty
  * Requires at least: 5.4
- * Tested up to: 5.8.1
+ * Tested up to: 5.9.1
  * Requires PHP: 7.4
  * phpcs:disable SlevomatCodingStandard.Namespaces.FullyQualifiedGlobalConstants.NonFullyQualified
  * @package TheFrosty\WpUpgradeTaskRunner
@@ -19,13 +19,14 @@
 namespace TheFrosty\WpUpgradeTaskRunner;
 
 const SLUG = 'wp-upgrade-task-runner';
-const VERSION = '2.5.0';
+const VERSION = '2.6.0';
 
 use TheFrosty\WpUpgradeTaskRunner\Cli\DispatchTasks;
 use TheFrosty\WpUpgradeTaskRunner\Upgrade\DbUpgrade;
 use TheFrosty\WpUpgradeTaskRunner\Upgrade\TaskCountCheck;
 use TheFrosty\WpUtilities\Plugin\PluginFactory;
 use TheFrosty\WpUtilities\WpAdmin\DisablePluginUpdateCheck;
+use function add_action;
 
 $plugin = PluginFactory::create(SLUG);
 /** Container object. @var \TheFrosty\WpUtilities\Plugin\Container $container */
@@ -47,6 +48,6 @@ $plugin
     ->addOnHook(DbUpgrade::class, 'admin_menu', null, true)
     ->addOnHook(TaskCountCheck::class, 'admin_menu', null, true);
 
-\add_action('plugins_loaded', static function () use ($plugin): void {
+add_action('plugins_loaded', static function () use ($plugin): void {
     $plugin->initialize();
 }, 5);
